The Great Storage Refactoring, Step 3: Make tests pass
